requirements-2.txt
beautifulsoup4
duckdb
fastapi
fitz
Flask
google_api_python_client
gradio
grpcio
matplotlib
nemo_toolkit
neo4j
nltk
nvidia_riva_client
openpyxl
pandas
passlib
Pillow
pydantic
pytesseract
ray
Requests
riva_api
scikit_learn
sounddevice
SpeechRecognition
SQLAlchemy
starlette
streamlit
torch
transformers
uvicorn
wikipedia
wordcloud
yt_dlp
psycopg2-binary
langchain